remove clippy check

+++ programs/cardinal-reward-distributor/src/instructions/init_reward_distributor.rs
@@ -48,7 +48,7 @@ pub fn handler<'key, 'accounts, 'remaining, 'info>(ctx: Context<'key, 'accounts,
     reward_distributor.stake_pool = ctx.accounts.stake_pool.key();
     reward_distributor.reward_mint = ctx.accounts.reward_mint.key();
     reward_distributor.reward_amount = ix.reward_amount;
-    reward_distributor.reward_duration_seconds = ix.reward_duration_seconds as u128;
+    reward_distributor.reward_duration_seconds = ix.reward_duration_seconds;
     reward_distributor.max_supply = ix.max_supply;
     reward_distributor.default_multiplier = ix.default_multiplier.unwrap_or(1);
     reward_distributor.multiplier_decimals = ix.multiplier_decimals.unwrap_or(0);
